+++ app/(minimal)/new/page.tsx
@@ -87,29 +87,21 @@ export default async function Page({ searchParams }: Props) {
   // text (with or without image)
   if (searchParams.rawText) {
     return (
-      <div className="flex w-full flex-col items-center">
-        {/* <YourDetails lists={lists || undefined} /> */}
-        {/* <ImageUpload filePath={searchParams.filePath} /> */}
-        <header className="fixed inset-x-0 top-2 flex items-center justify-center">
-          <ImagePreview filePath={searchParams.filePath} />
-        </header>
-        <Stages
-          lists={lists || undefined}
-          Preview={
-            <Suspense
-              fallback={<EventPreviewLoadingSpinner className="size-screen" />}
-            >
-              <EventsFromRawText
-                timezone={timezone}
-                rawText={searchParams.rawText}
-                filePath={searchParams.filePath}
-                edit={searchParams.edit}
-              />
-            </Suspense>
-          }
-        ></Stages>
-        <NewEventFooterButtons />
-      </div>
+      <Stages
+        lists={lists || undefined}
+        Preview={
+          <Suspense
+            fallback={<EventPreviewLoadingSpinner className="size-screen" />}
+          >
+            <EventsFromRawText
+              timezone={timezone}
+              rawText={searchParams.rawText}
+              filePath={searchParams.filePath}
+              edit={searchParams.edit}
+            />
+          </Suspense>
+        }
+      ></Stages>
     );
   }
 
@@ -119,31 +111,5 @@ export default async function Page({ searchParams }: Props) {
       <YourDetails lists={lists || undefined} />
       <AddEvent />
     </div>
-  );
-}
-
-function ImagePreview({ filePath }: { filePath?: string }) {
-  if (!filePath) return null;
-  const croppedImageUrl = Bytescale.UrlBuilder.url({
-    accountId: "12a1yek",
-    filePath: filePath, // Ensure filePath is defined and contains the path to the image
-    options: {
-      transformation: "image",
-      transformationParams: {
-        w: 48,
-        h: 48,
-        fit: "min",
-        q: 70,
-        f: "jpg",
-      },
-    },
-  });
-  return (
-    // eslint-disable-next-line @next/next/no-img-element
-    <img
-      src={croppedImageUrl}
-      alt="Image preview"
-      className="size-12 rounded-full object-cover"
-    />
   );
 }
